requirements.txt
python-dotenv
langchain
langchain-community
langchain-cohere
langchain-openai
flask
flask-cors